+++ Malmo/src/GoWrapper/AgentHost.go
@@ -23,7 +23,7 @@ package malmo
 #cgo CXXFLAGS: -I. -I.. -I../../../Schemas -std=c++11 -Wno-deprecated-declarations
 #cgo LDFLAGS: -L../../../build/install/Cpp_Examples/lib -lMalmo -lboost_system -lboost_filesystem -lboost_thread -lboost_iostreams -lboost_program_options -lboost_date_time -lboost_regex -lxerces-c
 
-#include "go_agenthost.h"
+#include "x_agent_host.h"
 #include "auxiliary.h"
 */
 import "C"
