+++ Malmo/src/GoWrapper/x_client_info.cpp
@@ -18,20 +18,20 @@
 // --------------------------------------------------------------------------------------------------
 
 // Malmo:
-#include <ClientPool.h>
+#include <ClientInfo.h>
 using namespace malmo;
 
 // Local:
-#include "go_clientpool.h"
+#include "x_client_info.h"
 
-ptClientPool new_client_pool() {
-    ClientPool * pt = new ClientPool;
+ptClientInfo new_client_info() {
+    ClientInfo * pt = new ClientInfo;
     return (void*)pt;
 }
 
-void free_client_pool(ptClientPool client_pool) {
-    if (client_pool != NULL) {
-        ClientPool * pt = (ClientPool*)client_pool;
+void free_client_info(ptClientInfo client_info) {
+    if (client_info != NULL) {
+        ClientInfo * pt = (ClientInfo*)client_info;
         delete pt;
     }
 }
