+++ Malmo/src/GoWrapper/x_client_info.h
