+++ Malmo/src/GoWrapper/x_client_pool.h
