+++ Malmo/src/GoWrapper/x_mission_record_spec.h
