+++ Malmo/src/GoWrapper/x_mission_spec.h
